+++ dsmigrator/antimalware.py
@@ -1,6 +1,15 @@
-from dsmigrator.AMconfigtenant1 import AMconfigtenant1
-from dsmigrator.CreateandReplaceAMConfigTenant2 import AmconfigTenant2, AmReplaceConfig
-from dsmigrator.AMCheckandRename import AmConfigCheck, RenameAmConfig
+import sys
+import os
+import time
+from time import sleep
+import requests
+import urllib3
+import json
+from dsmigrator.api_config import AMConfigApiInstance
+from dsmigrator.migrator_utils import validate_create
+
+
+cert = False
 
 
 def am_config_transform(antimalwareconfig, OLD_HOST, OLD_API_KEY):
@@ -44,3 +53,311 @@ def am_validate_create(
     )
     final = aop_replace_am_configs
     return final
+
+
+def AMconfigtenant1(antimalwareconfig, url_link_final, tenant1key):
+    # Describe each AM config from tenant1
+    allamconfig = []
+    directorylist = []
+    fileextentionlist = []
+    filelist = []
+    print("Getting Anti-Malware configuration from Tenant 1", flush=True)
+    for count, amconfig in enumerate(antimalwareconfig):
+        if int(amconfig) != 0:
+            payload = {}
+            url = url_link_final + "api/antimalwareconfigurations/" + str(amconfig)
+            headers = {
+                "api-secret-key": tenant1key,
+                "api-version": "v1",
+                "Content-Type": "application/json",
+            }
+            response = requests.request(
+                "GET", url, headers=headers, data=payload, verify=cert
+            )
+
+            describe = str(response.text)
+            allamconfig.append(describe)
+            index = describe.find("directoryListID")
+            if index != -1:
+                indexpart = describe[index + 16 :]
+                startIndex = indexpart.find(":")
+                if startIndex != -1:  # i.e. if the first quote was found
+                    endIndex = indexpart.find(",", startIndex + 1)
+                    if (
+                        startIndex != -1 and endIndex != -1
+                    ):  # i.e. both quotes were found
+                        indexid = indexpart[startIndex + 1 : endIndex]
+                        directorylist.append(str(indexid))
+            index = describe.find("excludedDirectoryListID")
+            if index != -1:
+                indexpart = describe[index + 24 :]
+                startIndex = indexpart.find(":")
+                if startIndex != -1:  # i.e. if the first quote was found
+                    endIndex = indexpart.find(",", startIndex + 1)
+                    if (
+                        startIndex != -1 and endIndex != -1
+                    ):  # i.e. both quotes were found
+                        indexid = indexpart[startIndex + 1 : endIndex]
+                        directorylist.append(str(indexid))
+            index = describe.find("excludedFileExtensionListID")
+            if index != -1:
+                indexpart = describe[index + 28 :]
+                startIndex = indexpart.find(":")
+                if startIndex != -1:  # i.e. if the first quote was found
+                    endIndex = indexpart.find(",", startIndex + 1)
+                    if (
+                        startIndex != -1 and endIndex != -1
+                    ):  # i.e. both quotes were found
+                        indexid = indexpart[startIndex + 1 : endIndex]
+                        fileextentionlist.append(str(indexid))
+            index = describe.find("fileExtensionListID")
+            if index != -1:
+                indexpart = describe[index + 20 :]
+                startIndex = indexpart.find(":")
+                if startIndex != -1:  # i.e. if the first quote was found
+                    endIndex = indexpart.find(",", startIndex + 1)
+                    if (
+                        startIndex != -1 and endIndex != -1
+                    ):  # i.e. both quotes were found
+                        indexid = indexpart[startIndex + 1 : endIndex]
+                        fileextentionlist.append(str(indexid))
+            index = describe.find("excludedFileListID")
+            if index != -1:
+                indexpart = describe[index + 19 :]
+                startIndex = indexpart.find(":")
+                if startIndex != -1:  # i.e. if the first quote was found
+                    endIndex = indexpart.find(",", startIndex + 1)
+                    if (
+                        startIndex != -1 and endIndex != -1
+                    ):  # i.e. both quotes were found
+                        indexid = indexpart[startIndex + 1 : endIndex]
+                        filelist.append(str(indexid))
+            index = describe.find("excludedProcessImageFileListID")
+            if index != -1:
+                indexpart = describe[index + 31 :]
+                startIndex = indexpart.find(":")
+                if startIndex != -1:  # i.e. if the first quote was found
+                    endIndex = indexpart.find(",", startIndex + 1)
+                    if (
+                        startIndex != -1 and endIndex != -1
+                    ):  # i.e. both quotes were found
+                        indexid = indexpart[startIndex + 1 : endIndex]
+                        filelist.append(str(indexid))
+            print(
+                "#" + str(count) + " Anti-Malware Config ID: " + str(amconfig),
+                flush=True,
+            )
+    directorylist = list(dict.fromkeys(directorylist))
+    fileextentionlist = list(dict.fromkeys(fileextentionlist))
+    filelist = list(dict.fromkeys(filelist))
+    return directorylist, fileextentionlist, filelist, allamconfig
+
+
+def AmconfigTenant2(allamconfig):
+    allamconfignew = []
+    print("Creating Anti-Malware Configuration to Tenant2", flush=True)
+    if allamconfig:
+        validate_create(allamconfig, AMConfigApiInstance(), "anti malware")
+    print("New AM Config ID", flush=True)
+    print(allamconfignew, flush=True)
+    return allamconfignew
+
+
+def AmReplaceConfig(allofpolicy, antimalwareconfig, allamconfignew):
+    count = 0
+    for describe in allofpolicy:
+        count1 = 0
+        index = describe.find("realTimeScanConfigurationID")
+        if index != -1:
+            indexpart = describe[index + 28 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in antimalwareconfig:
+                        if int(dirlist) != 0:
+                            if indexid == dirlist:
+                                describe = (
+                                    describe[: index + 28 + startIndex + 1]
+                                    + allamconfignew[count1]
+                                    + describe[index + 28 + startIndex + endIndex :]
+                                )
+                            count1 = count1 + 1
+        count1 = 0
+        index = describe.find("manualScanConfigurationID")
+        if index != -1:
+            indexpart = describe[index + 26 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in antimalwareconfig:
+                        if int(dirlist) != 0:
+                            if indexid == dirlist:
+                                describe = (
+                                    describe[: index + 26 + startIndex + 1]
+                                    + allamconfignew[count1]
+                                    + describe[index + 26 + startIndex + endIndex :]
+                                )
+                            count1 = count1 + 1
+        count1 = 0
+        index = describe.find("scheduledScanConfigurationID")
+        if index != -1:
+            indexpart = describe[index + 29 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find("}", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in antimalwareconfig:
+                        if int(dirlist) != 0:
+                            if indexid == dirlist:
+                                describe = (
+                                    describe[: index + 29 + startIndex + 1]
+                                    + allamconfignew[count1]
+                                    + describe[index + 29 + startIndex + endIndex :]
+                                )
+                            count1 = count1 + 1
+
+        allofpolicy[count] = describe
+        count = count + 1
+    return allofpolicy
+
+
+def AmConfigCheck(
+    allamconfig,
+    directorylist,
+    alldirectorynew,
+    fileextentionlist,
+    allfileextentionnew,
+    filelist,
+    allfilelistnew,
+):
+
+    count = 0
+    for describe in allamconfig:
+        count1 = 0
+        index = describe.find('"directoryListID"')
+        if index != -1:
+            indexpart = describe[index + 17 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in directorylist:
+                        if indexid == dirlist:
+                            if alldirectorynew:
+                                describe = (
+                                    describe[: index + 17 + startIndex + 1]
+                                    + alldirectorynew[count1]
+                                    + describe[index + 17 + startIndex + endIndex :]
+                                )
+                        count1 = count1 + 1
+        count1 = 0
+        index = describe.find("excludedDirectoryListID")
+        if index != -1:
+            indexpart = describe[index + 24 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in directorylist:
+                        if indexid == dirlist:
+                            if alldirectorynew:
+                                describe = (
+                                    describe[: index + 24 + startIndex + 1]
+                                    + alldirectorynew[count1]
+                                    + describe[index + 24 + startIndex + endIndex :]
+                                )
+                        count1 = count1 + 1
+        count1 = 0
+        index = describe.find("excludedFileExtensionListID")
+        if index != -1:
+            indexpart = describe[index + 28 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in fileextentionlist:
+                        if indexid == dirlist:
+                            if allfileextentionnew:
+                                describe = (
+                                    describe[: index + 28 + startIndex + 1]
+                                    + allfileextentionnew[count1]
+                                    + describe[index + 28 + startIndex + endIndex :]
+                                )
+                        count1 = count1 + 1
+        count1 = 0
+        index = describe.find('"fileExtensionListID"')
+        if index != -1:
+            indexpart = describe[index + 21]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in fileextentionlist:
+                        if indexid == dirlist:
+                            if allfileextentionnew:
+                                describe = (
+                                    describe[: index + 21 + startIndex + 1]
+                                    + allfileextentionnew[count1]
+                                    + describe[index + 21 + startIndex + endIndex :]
+                                )
+                        count1 = count1 + 1
+        count1 = 0
+        index = describe.find("excludedFileListID")
+        if index != -1:
+            indexpart = describe[index + 19 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in filelist:
+                        if indexid == dirlist:
+                            if allfilelistnew:
+                                describe = (
+                                    describe[: index + 19 + startIndex + 1]
+                                    + allfilelistnew[count1]
+                                    + describe[index + 19 + startIndex + endIndex :]
+                                )
+                        count1 = count1 + 1
+        count1 = 0
+        index = describe.find("excludedProcessImageFileListID")
+        if index != -1:
+            indexpart = describe[index + 31 :]
+            startIndex = indexpart.find(":")
+            if startIndex != -1:  # i.e. if the first quote was found
+                endIndex = indexpart.find(",", startIndex + 1)
+                if startIndex != -1 and endIndex != -1:  # i.e. both quotes were found
+                    indexid = indexpart[startIndex + 1 : endIndex]
+                    for dirlist in filelist:
+                        if indexid == dirlist:
+                            if allfilelistnew:
+                                describe = (
+                                    describe[: index + 31 + startIndex + 1]
+                                    + allfilelistnew[count1]
+                                    + describe[index + 31 + startIndex + endIndex :]
+                                )
+                        count1 = count1 + 1
+        allamconfig[count] = describe
+        count = count + 1
+    return allamconfig
+
+
+def RenameAmConfig(allamconfig):
+    count = 0
+    if allamconfig:
+        for describe in allamconfig:
+            amjson = json.loads(describe)
+            print(describe)
+            amjson["name"] = amjson["name"] + " - Migrated"
+            allamconfig[count] = json.dumps(amjson)
+            count = count + 1
+    return allamconfig
